a2_Language_Model/app/requirements.txt
Flask==2.3.3
scipy
matplotlib
numpy
datasets==3.2.0
torchdata==0.6.0
torchtext==0.15.0